+++ src/rotation.c
@@ -1,43 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   transform.c                                        :+:      :+:    :+:   */
+/*   rotation.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nde-sant <nde-sant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/11/12 16:36:27 by nde-sant          #+#    #+#             */
-/*   Updated: 2025/11/20 13:33:37 by nde-sant         ###   ########.fr       */
+/*   Created: 2025/11/27 15:49:40 by nde-sant          #+#    #+#             */
+/*   Updated: 2025/11/27 15:50:13 by nde-sant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
-
-void	scale_transform(t_point *points, t_vector scalar, int size)
-{
-	int	i;
-
-	i = 0;
-	while (i < size)
-	{
-		points[i].x *= scalar.x;
-		points[i].y *= scalar.y;
-		points[i].z *= scalar.z;
-		i++;
-	}
-}
-
-void	translate_2d(t_point *points, int size, int dx, int dy)
-{
-	int	i;
-
-	i = 0;
-	while (i < size)
-	{
-		points[i].x += dx;
-		points[i].y += dy;
-		i++;
-	}
-}
 
 void	rotate_x_axis(t_point *pts, double angle, int size)
 {
